redo file layout so that all components are located in components
