+++ client/src/modules/Admin/Components/AdminUser.tsx
@@ -7,6 +7,13 @@ type Props = {
   removeHandler: (arg1: any) => any;
 };
 
+/**
+ * AdminUser Component
+ *
+ * Displays an individual admin user entry, including their first name,
+ * last name, NetID, and a button to remove the admin from the system.
+ */
+
 const AdminUser = ({ user, token, removeHandler }: Props) => {
   return (
     <div className={styles.userEntry}>
