--- client/src/modules/Admin/Pages/ManageAdmins/ManageAdmins.module.css
@@ -1,15 +1,0 @@
-.userEntry {
-  padding: 1px;
-  margin: 2px;
-  display: flex;
-  border-bottom: 1px solid #ccc;
-}
-
-.removeButton {
-  color: var(--clr-black);
-  font-size: var(--font-small-size);
-  border-radius: 5px;
-  margin-left: auto;
-  padding: 2px 3px;
-  cursor: pointer;
-}
